ENT-3482: Move BC crypto service implementation to node api.
